fix: rename to JobsList
